requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
faiss-cpu==1.8.0
pypdf==4.3.1
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.7.4
python-multipart==0.0.9